update chrony section
The way the exercise is written implies that the NTP service must be enabled on the firewall for chrony to work.

With this change, I try to make it more clear that chrony can act as a client or server, and that the firewall port need only be opened if running as a server.